+++ NEWS.md
@@ -1,6 +1,7 @@
 # idaifieldR 0.3.7 _2026-03-XX_
 ## New Field Desktop API
 * idaifieldR now gets the configuration of a project from Field Desktop directly via the `/config/` enpoint instead of the database document. That means that the complete configuration is now accessible.
+  * `get_field_inputtypes()` produces a data.frame of Categories, Parent Categories, input field identifiers and the corresponding `inputType` in the given Project Configuration.
 
 ## Renewed `simplify_idaifield()`
 `simplify_idaifield()` has been drastically changed. Some functionality which would have created data issues for users was removed (e.g. dimension-recalculation), which results in a less "simple" return list, but will keep the data intact. The function is now less verbose. All lists like `campaign` or `processor` or any checkbox-fields that can contain multiple values are returned as vectors instead of lists. Geometry will be parsed to JSON if kept. UUIDs are replaced during processing.  
@@ -20,7 +21,6 @@
 * Removed `reformat_geometry()`: Instead, I recommend (and do) re-json-ify the Geometry and advise to use `sf::st_read(json_string, quiet = TRUE)` on each geometry elsewhere.
 * Removed `get_language_lookup()` and removed language support from `simplify_idaifield()` - it's just too much.
 * Removed `extract_field_names()`
-* Replaced `get_field_inputtypes()` with (experimental) `extract_inputtypes()`.
 * Reworked `check_and_unnest()` into `maybe_unnest_docs()`.
 * Reworked `find_layers()` to require a vector of categories to search for as layer_categories instead of a global option.